+++ app/src/main/java/com/vivifram/second/hitalk/ui/view/CircleProgressBar.java
@@ -197,7 +197,7 @@ public class CircleProgressBar extends ImageView {
                 ViewCompat.setLayerType(this, ViewCompat.LAYER_TYPE_SOFTWARE, mBgCircle.getPaint());
                 mBgCircle.getPaint().setShadowLayer(mShadowRadius, shadowXOffset, shadowYOffset,
                         KEY_SHADOW_COLOR);
-                final int padding = (int) mShadowRadius;
+                final int padding = mShadowRadius;
                 // set padding so the inner image sits correctly within the shadow.
                 setPadding(padding, padding, padding, padding);
             }
